+++ src/View/Helper/AssetHelper.php
@@ -35,10 +35,10 @@ class AssetHelper extends Helper
      * @param string $type `css` or `js`
      * @return string Asset path
      * @uses Assets\Utility\AssetsCreator::create()
-     * @uses Assets\Utility\AssetsCreator::getAssetFilename()
-     * @uses Assets\Utility\AssetsCreator::getAssetPath()
+     * @uses Assets\Utility\AssetsCreator::filename()
+     * @uses Assets\Utility\AssetsCreator::path()
      */
-    protected function getAssetPath($path, $type)
+    protected function path($path, $type)
     {
         if (Configure::read('debug') && !Configure::read(ASSETS . '.force')) {
             return $path;
@@ -46,13 +46,13 @@ class AssetHelper extends Helper
 
         $asset = new AssetsCreator($path, $type);
         $asset->create();
-        $path = '/assets/' . $asset->getAssetFilename();
+        $path = '/assets/' . $asset->filename();
 
         //Appends the timestamp
         $stamp = Configure::read('Asset.timestamp');
         $timestampEnabled = $stamp === 'force' || ($stamp === true && Configure::read('debug'));
         if ($timestampEnabled) {
-            $path .= '.' . $type . '?' . filemtime($asset->getAssetPath());
+            $path .= '.' . $type . '?' . filemtime($asset->path());
         }
 
         return $path;
@@ -63,11 +63,11 @@ class AssetHelper extends Helper
      * @param string|array $path String or array of css files
      * @param array $options Array of options and HTML attributes
      * @return string Html, `<link>` or `<style>` tag
-     * @uses getAssetPath()
+     * @uses path()
      */
     public function css($path, array $options = [])
     {
-        $path = $this->getAssetPath($path, 'css');
+        $path = $this->path($path, 'css');
 
         return $this->Html->css($path, $options);
     }
@@ -78,11 +78,11 @@ class AssetHelper extends Helper
      * @param array $options Array of options and HTML attributes
      * @return mixed String of `<script />` tags or null if `$inline` is
      *  false or if `$once` is true
-     * @uses getAssetPath()
+     * @uses path()
      */
     public function script($url, array $options = [])
     {
-        $url = $this->getAssetPath($url, 'js');
+        $url = $this->path($url, 'js');
 
         return $this->Html->script($url, $options);
     }